+++ src/jbsnorro/HashSet{T}.ts
