+++ src/jbsnorro/polygons/ManhattanToBoundaryComparer.ts
