+++ src/jbsnorro/polygons/MinDistances.ts
