+++ src/jbsnorro/polygons/RectanglesToPolygon.ts
